Rag_bot/requirements.txt
pypdf
openai
pinecone
python-dotenv
google-generativeai
streamlit
langchain
